+++ src/StackAdmin/Network/Commands.Network.Test/ScenarioTests/VirtualNetworkTests.ps1
@@ -33,8 +33,10 @@ function Test-VirtualNetworkCRUD
         
         # Create the Virtual Network
         $subnet = New-AzureRmVirtualNetworkSubnetConfig -Name $subnetName -AddressPrefix 10.0.1.0/24
-        $actual = New-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -DnsServer 8.8.8.8 -Subnet $subnet
-        $expected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+        $job = New-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -DnsServer 8.8.8.8 -Subnet $subnet -AsJob
+        $job | Wait-Job
+		$actual = $job | Receive-Job
+		$expected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
         
         Assert-AreEqual $expected.ResourceGroupName $rgname    
         Assert-AreEqual $expected.Name $actual.Name    
@@ -80,8 +82,10 @@ function Test-VirtualNetworkCRUD
         Assert-AreEqual 5 @($testResponse2.AvailableIpAddresses).Count
         
         # Delete VirtualNetwork
-        $delete = Remove-AzureRmvirtualNetwork -ResourceGroupName $rgname -name $vnetName -PassThru -Force
-        Assert-AreEqual true $delete
+        $job = Remove-AzureRmvirtualNetwork -ResourceGroupName $rgname -name $vnetName -PassThru -Force -AsJob
+        $job | Wait-Job
+		$delete = $job | Receive-Job
+		Assert-AreEqual true $delete
                 
         $list = Get-AzureRmvirtualNetwork -ResourceGroupName $rgname
         Assert-AreEqual 0 @($list).Count
@@ -134,8 +138,9 @@ function Test-subnetCRUD
         Assert-AreEqual "10.0.2.0/24" $vnetExpected.Subnets[1].AddressPrefix
         
         # Edit a subnet
-        Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname | Set-AzureRmVirtualNetworkSubnetConfig -Name $subnet2Name -AddressPrefix 10.0.3.0/24 | Set-AzureRmVirtualNetwork
-        
+        $job = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname | Set-AzureRmVirtualNetworkSubnetConfig -Name $subnet2Name -AddressPrefix 10.0.3.0/24 | Set-AzureRmVirtualNetwork -AsJob
+        $job | Wait-Job
+
         $vnetExpected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
         Assert-AreEqual 2 @($vnetExpected.Subnets).Count
         Assert-AreEqual $subnetName $vnetExpected.Subnets[0].Name
@@ -157,6 +162,67 @@ function Test-subnetCRUD
         $vnetExpected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
         Assert-AreEqual 1 @($vnetExpected.Subnets).Count
         Assert-AreEqual $subnetName $vnetExpected.Subnets[0].Name        
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Tests creating new simple virtualNetwork with DDoSProtecion parameters.
+#>
+function Test-VirtualNetworkCRUDWithDDoSProtection
+{
+    # Setup
+    $rgname = Get-ResourceGroupName
+    $vnetName = Get-ResourceName
+    $subnetName = Get-ResourceName
+    $rglocation = Get-ProviderLocation ResourceManagement
+    $resourceTypeParent = "Microsoft.Network/virtualNetworks"
+    $location = Get-ProviderLocation $resourceTypeParent
+    
+    try 
+    {
+        # Create the resource group
+        $resourceGroup = New-AzureRmResourceGroup -Name $rgname -Location $rglocation -Tags @{ testtag = "testval" } 
+        
+        # Create the Virtual Network
+        $subnet = New-AzureRmVirtualNetworkSubnetConfig -Name $subnetName -AddressPrefix 10.0.1.0/24
+        $actual = New-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -DnsServer 8.8.8.8 -Subnet $subnet -EnableDDoSProtection
+        $expected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+        
+        Assert-AreEqual $expected.ResourceGroupName $rgname    
+        Assert-AreEqual $expected.Name $actual.Name    
+        Assert-AreEqual $expected.Location $actual.Location
+        Assert-AreEqual "Succeeded" $expected.ProvisioningState
+        Assert-NotNull $expected.ResourceGuid
+        Assert-AreEqual "10.0.0.0/16" $expected.AddressSpace.AddressPrefixes[0]
+        Assert-AreEqual 1 @($expected.DhcpOptions.DnsServers).Count
+        Assert-AreEqual "8.8.8.8" $expected.DhcpOptions.DnsServers[0]
+        Assert-AreEqual 1 @($expected.Subnets).Count
+        Assert-AreEqual $subnetName $expected.Subnets[0].Name
+        Assert-AreEqual "10.0.1.0/24" $expected.Subnets[0].AddressPrefix
+		Assert-AreEqual true $expected.EnableDDoSProtection
+		Assert-AreEqual false $expected.EnableVmProtection
+        
+		$expected.EnableDDoSProtection=$false
+		Set-AzureRmVirtualNetwork -VirtualNetwork $expected
+		$expected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+        Assert-AreEqual false $expected.EnableDDoSProtection
+		Assert-AreEqual false $expected.EnableVmProtection
+
+		$expected.EnableVmProtection=$true
+		Set-AzureRmVirtualNetwork -VirtualNetwork $expected
+		$expected = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+        Assert-AreEqual false $expected.EnableDDoSProtection
+		Assert-AreEqual true $expected.EnableVmProtection
+        
+        # Delete VirtualNetwork
+        $delete = Remove-AzureRmvirtualNetwork -ResourceGroupName $rgname -name $vnetName -PassThru -Force
+        Assert-AreEqual true $delete
     }
     finally
     {
@@ -207,7 +273,9 @@ function Test-VirtualNetworkPeeringCRUD
         Assert-AreEqual "Succeeded" $vnet2.ProvisioningState 
 
         # Add Peering to vnet1
-        $peer = $vnet1 | Add-AzureRmVirtualNetworkPeering -name $peerName -RemoteVirtualNetworkId $vnet2.Id -AllowForwardedTraffic
+        $job = $vnet1 | Add-AzureRmVirtualNetworkPeering -name $peerName -RemoteVirtualNetworkId $vnet2.Id -AllowForwardedTraffic -AsJob
+		$job | Wait-Job
+		$peer = $job | Receive-Job
         
         Assert-AreEqual $peer.ResourceGroupName $rgname    
         Assert-AreEqual $peer.Name $peerName    
@@ -253,7 +321,9 @@ function Test-VirtualNetworkPeeringCRUD
         # Set Peer
         $getPeer.AllowForwardedTraffic = $false
         
-        $setPeer = $getPeer | Set-AzureRmVirtualNetworkPeering
+        $job = $getPeer | Set-AzureRmVirtualNetworkPeering -AsJob
+		$job | Wait-Job
+		$setPeer = $job | Receive-Job
         
         Assert-AreEqual $setPeer.ResourceGroupName $rgname    
         Assert-AreEqual $setPeer.Name $peerName    
@@ -268,7 +338,9 @@ function Test-VirtualNetworkPeeringCRUD
         Assert-Null $setPeer.$peer.RemoteVirtualNetworkAddressSpace
         
         # Delete Peer
-        $delete = Remove-AzureRmVirtualNetworkPeering -name $peerName -VirtualNetworkName $vnet1Name -ResourceGroupName $rgname -Force -PassThru
+        $job = Remove-AzureRmVirtualNetworkPeering -name $peerName -VirtualNetworkName $vnet1Name -ResourceGroupName $rgname -Force -PassThru -AsJob
+		$job | Wait-Job
+		$delete = $job | Receive-Job
         Assert-AreEqual true $delete
 
         # Delete VirtualNetwork
@@ -393,6 +465,50 @@ function Test-VirtualNetworkUsage
         $currentUsageNew = $usage.CurrentValue;
 
         Assert-AreEqual $currentUsage $($currentUsageNew - 1);
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Tests checking Virtual Network Subnet Service Endpoint feature.
+#>
+function Test-VirtualNetworkSubnetServiceEndpoint
+{
+    # Setup
+    $rgname = Get-ResourceGroupName
+    $vnetName = Get-ResourceName
+    $subnetName = Get-ResourceName
+    $rglocation = Get-ProviderLocation ResourceManagement
+    $resourceTypeParent = "Microsoft.Network/virtualNetworks"
+    $location = Get-ProviderLocation $resourceTypeParent
+    $serviceEndpoint = "Microsoft.Storage"
+
+    try
+    {
+        # Create the resource group
+        $resourceGroup = New-AzureRmResourceGroup -Name $rgname -Location $rglocation -Tags @{ testtag = "testval" };
+
+        # Create the Virtual Network
+        $subnet = New-AzureRmVirtualNetworkSubnetConfig -Name $subnetName -AddressPrefix 10.0.1.0/24 -ServiceEndpoint $serviceEndpoint;
+        New-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -Subnet $subnet;
+        $vnet = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname;
+
+        Assert-NotNull $vnet;
+        Assert-NotNull $vnet.Subnets;
+
+        $subnet = $vnet.Subnets[0];
+        Assert-AreEqual $serviceEndpoint $subnet.serviceEndpoints[0].Service;
+
+        Set-AzureRmVirtualNetworkSubnetConfig -Name $subnetName -VirtualNetwork $vnet -AddressPrefix 10.0.1.0/24 -ServiceEndpoint $null;
+        $vnet = Set-AzureRmVirtualNetwork -VirtualNetwork $vnet;
+        $subnet = $vnet.Subnets[0];
+
+        Assert-Null $subnet.serviceEndpoints;
     }
     finally
     {